Start of HTML and Skeleton CSS
I am currently testing skeleton css and incorporating it into a mock-up of the html we will begin working on for the project. This html will most likely change but a simple start to test boundaries will help us move forward and see what style and functionality we want to build from.