Skeleton CSS and Very Basic Layout
I used Skeleton CSS to add a heading, title, image, buttons, and text area and the rest of the styling was done with vanilla. This sets us up to apply whatever content we want and if we need to build from it we can use the current terminology to do so.