Saving Meme/Quote Data to Local Storage
I created some functions based on our last assignment to create a method of saved hsitory on our page. Ideally we will have an unordered list display next to the presented memes and have them as links to click on to display a previous combination of meme and text. THis will be saved to local storage.