Quote Generator and Path to save code
I added the second api which is the quote generator and implemented it below the meme generator. I commented out the local storage functions because they may be applicable but I need to decide whether we need to save the page as a whole or each meme/quote individually to local storage.