Saving to Local Storage
I managed to make one button to save both current quote text and current meme image from the two separate apis save to local storage. Now I must figure out how to save a button to the side of the webpage with that stored data with a text box to name that data.